requirements.txt
py-trello==0.9.0
PyQt5==5.8.2
QtPy==1.2.1
pyqttango==0.1.2